Flat.Decoder

Description

Strict Decoder

Documentation

strictDecoder :: Get a -> ByteString -> Either DecodeException a

strictDecoderPart :: Get a -> ByteString -> Either DecodeException a

type Decoded a = Either DecodeException a

A decoded value

data DecodeException

An exception during decoding

data Get a

Decoder monad

dByteString :: Get ByteString

dLazyByteString :: Get ByteString

dShortByteString :: Get ShortByteString

dShortByteString_ :: Get ShortByteString

dUTF16 :: Get Text

dUTF8 :: Get Text

decodeArrayWith :: Get a -> Get [a]

decodeListWith :: Get a -> Get [a]

dFloat :: Get Float

Decode a Float

dDouble :: Get Double

Decode a Double

dInteger :: Get Integer

dNatural :: Get Natural

dChar :: Get Char

dBool :: Get Bool

Decode a boolean

dWord8 :: Get Word8

Return the 8 most significant bits (same as dBE8)

dWord16 :: Get Word16

dWord32 :: Get Word32

dWord64 :: Get Word64

dWord :: Get Word

dInt8 :: Get Int8

dInt16 :: Get Int16

dInt32 :: Get Int32

dInt64 :: Get Int64

dInt :: Get Int

dBE8 :: Get Word8

Return the 8 most significant bits

dBE16 :: Get Word16

Return the 16 most significant bits

dBE32 :: Get Word32

Return the 32 most significant bits

dBE64 :: Get Word64

Return the 64 most significant bits

dBEBits8 :: Int -> Get Word8

Return the n most significant bits (up to maximum of 8)

The bits are returned right shifted: >>> unflatWith (dBEBits8 3) [0b11100001::Word8] == Right 0b00000111 True

dBEBits16 :: Int -> Get Word16

Return the n most significant bits (up to maximum of 16) The bits are returned right shifted.

dBEBits32 :: Int -> Get Word32

Return the n most significant bits (up to maximum of 8) The bits are returned right shifted.

dBEBits64 :: Int -> Get Word64

Return the n most significant bits (up to maximum of 8) The bits are returned right shifted.

dropBits :: Int -> Get ()

Drop the specified number of bits

data ConsState

A special state, optimised for constructor decoding. It consists of: The bits to parse, top bit being the first to parse (could use a Word16 instead, no difference in performance) The number of decoded bits Supports up to 512 constructors (9 bits)

Constructors

ConsState !Word !Int

consOpen :: Get ConsState

Switch to constructor decoding {--}

consClose :: Int -> Get ()

Switch back to normal decoding {--}

consBool :: ConsState -> (ConsState, Bool)

Decode a single bit

consBits :: ConsState -> Int -> (ConsState, Word)

Decode from 1 to 3 bits It could read more bits that are available, but it doesn't matter, errors will be checked in consClose