Flat.Encoder.Prim

Description

Encoding Primitives

Documentation

eBits16F :: NumBits -> Word16 -> Prim

Encode up to 9 bits

eBitsF :: NumBits -> Word8 -> Prim

Encode up to 8 bits.

eFloatF :: Float -> Prim

eDoubleF :: Double -> Prim

eUTF16F :: Text -> Prim

Encode text as UTF16 and encode the result as an array of bytes Efficient, as Text is already internally encoded as UTF16.

eUTF8F :: Text -> Prim

Encode text as UTF8 and encode the result as an array of bytes PROB: encodeUtf8 calls a C primitive, not compatible with GHCJS

eCharF :: Char -> Prim

eNaturalF :: Natural -> Prim

eIntegerF :: Integer -> Prim

eInt64F :: Int64 -> Prim

eInt32F :: Int32 -> Prim

eIntF :: Int -> Prim

eInt16F :: Int16 -> Prim

eInt8F :: Int8 -> Prim

eWordF :: Word -> Prim

eWord64F :: Word64 -> Prim

eWord32F :: Word32 -> Prim

eWord16F :: Word16 -> Prim

eBytesF :: ByteString -> Prim

eLazyBytesF :: ByteString -> Prim

eShortBytesF :: ShortByteString -> Prim

eWord8F :: Word8 -> Prim

eFillerF :: Prim

eBoolF :: Bool -> Prim

eTrueF :: Prim

eFalseF :: Prim

varWordF :: (Bits t, Integral t) => t -> Prim

w7l :: (Bits t, Integral t) => t -> [Word8]

Exported for testing only

eWord32BEF :: Word32 -> Prim

eWord64BEF :: Word64 -> Prim

eWord32E :: (Word32 -> Word32) -> Word32 -> Prim

eWord64E :: (Word64 -> Word64) -> Word64 -> Prim