Flat.Types

Description

Common Types

Documentation

type NumBits = Int

Number of bits

module Data.Word

module Data.Int

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data ShortByteString

A compact representation of a Word8 vector.

It has a lower memory overhead than a ByteString and does not contribute to heap fragmentation. It can be converted to or from a ByteString (at the cost of copying the string data). It supports very few other operations.

It is suitable for use as an internal representation for code that needs to keep many short strings in memory, but it should not be used as an interchange type. That is, it should not generally be used in public APIs. The ByteString type is usually more suitable for use in interfaces; it is more flexible and it supports a wide range of operations.

data Text

A space efficient, packed, unboxed Unicode text type.